BUG: Fix rounding to 0 errors and flat region errors
Rounding to 0 errors: use float datatype instead of int64

Flat region errors: occurs in getHistogram. This occurs as only one edge is calculated, which causes all values to be digitized to '0' in `binImage`, where the code expects the values to start a 1 after binning.
Fix: If a flat region is present, ensure that image is 'binned' using 1 bin, which causes binImage to digitize to '1'.

+++ radiomics/base.py
@@ -35,7 +35,7 @@ class RadiomicsFeaturesBase(object):
     self.imageArray = sitk.GetArrayFromImage(self.inputImage)
     self.maskArray = (sitk.GetArrayFromImage(self.inputMask) == self.label).astype('int')
 
-    self.matrix = self.imageArray.astype('int64')
+    self.matrix = self.imageArray.astype('float')
     self.matrixCoordinates = numpy.where(self.maskArray != 0)
 
     self.targetVoxelArray = self.matrix[self.matrixCoordinates]
